+++ src/octowrap/__init__.py
@@ -1,3 +1,3 @@
 """Rewrap Python # comments to a specified line length."""
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"